% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{add_optparse_options}
\alias{add_optparse_options}
\title{Add optparse options to a parser}
\usage{
add_optparse_options(parser, par_set)
}
\arguments{
\item{parser}{An optparse parser}

\item{par_set}{A parameter set}
}
\description{
Add optparse options to a parser
}
