% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_super.R
\name{argparse_trafo}
\alias{argparse_trafo}
\title{Parse a character representation to a value of the correct type}
\usage{
argparse_trafo(x, v)
}
\arguments{
\item{x}{A parameter object}

\item{v}{A character representation of a parameter value}
}
\description{
Parse a character representation to a value of the correct type
}
