% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_finite.R
\name{check_finite}
\alias{check_finite}
\title{Helper function for checking whether a vector is finite and of length one}
\usage{
check_finite(x)
}
\arguments{
\item{x}{Vector to check for length 1 finiteness}
}
\description{
Helper function for checking whether a vector is finite and of length one
}
