% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_collapse_set.R
\name{collapse_set}
\alias{collapse_set}
\title{A helper function for collapsing a set}
\usage{
collapse_set(..., sep = ", ", prefix = "{", postfix = "}")
}
\arguments{
\item{...}{Characters to collapse}

\item{sep}{Seperator between elements}

\item{prefix}{A prefix}

\item{postfix}{A postfix}
}
\description{
Will surround the collapsed set with brackets if it has more than one element.
}
