% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_collapse_vector.R
\name{collapse_vector}
\alias{collapse_vector}
\title{A helper function for collapsing a vector}
\usage{
collapse_vector(vec)
}
\description{
Will surround the collapsed vector with brackets if it has more than one element.
}
