% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_super.R
\name{distribution_function}
\alias{distribution_function}
\alias{quantile_function}
\alias{list_as_distribution}
\title{Helper functions for converting distributions from and to other formats}
\usage{
distribution_function(dist)

quantile_function(dist)

list_as_distribution(li)
}
\arguments{
\item{dist}{A distribution object.}

\item{li}{A list to be converted into a distribution.}
}
\description{
Helper functions for converting distributions from and to other formats
}
\seealso{
\link{dynparam}.
}
