% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_expuniform.R
\name{expuniform_distribution}
\alias{expuniform_distribution}
\title{Exponentially scaled uniform distribution.}
\usage{
expuniform_distribution(lower, upper)
}
\arguments{
\item{lower}{Lower limit of the distribution.}

\item{upper}{Upper limit of the distribution.}
}
\description{
Distributions are used for defining the domain of an \code{\link[=integer_parameter]{integer_parameter()}} or \code{\link[=numeric_parameter]{numeric_parameter()}}.
}
\examples{
expuniform_distribution(1, 10000)

expuniform_distribution(1e-5, 1e-2)
}
\seealso{
\link{dynparam}
}
