% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_super.R
\name{get_description}
\alias{get_description}
\title{Get a description of the parameter}
\usage{
get_description(x, sep = ", ")
}
\arguments{
\item{x}{The parameter}

\item{sep}{A separator between different fields}
}
\description{
Get a description of the parameter
}
