% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_super.R
\name{as_paramhelper}
\alias{as_paramhelper}
\alias{as_roxygen}
\alias{as_argparse}
\alias{as_list}
\alias{list_as_parameter}
\title{Helper functions for converting parameters from and to other formats}
\usage{
as_paramhelper(param)

as_roxygen(param)

as_argparse(param)

as_list(x)

list_as_parameter(li)
}
\arguments{
\item{param}{A parameter to be converted.}

\item{x}{An object (parameter or distribution) to be converted.}

\item{li}{A list to be converted into a parameter.}
}
\description{
Helper functions for converting parameters from and to other formats
}
\seealso{
\link{dynparam}
}
