% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_range.R
\name{range_parameter}
\alias{range_parameter}
\title{Define a range parameter}
\usage{
range_parameter(id, lower_default, upper_default, lower_distribution,
  upper_distribution, description = NULL)
}
\arguments{
\item{id}{The name of the parameter.}

\item{lower_default}{The default lower value of the range.}

\item{upper_default}{The default upper value of the range.}

\item{lower_distribution}{A distribution from which the lower value of the range can be sampled.
See \link[dynparam:dynparam]{?dynparam} for a list of possible distributions.}

\item{upper_distribution}{A distribution from which the upper value fo the range can be sampled.
See \link[dynparam:dynparam]{?dynparam} for a list of possible distributions.}

\item{description}{An optional (but recommended) description of the parameter.}
}
\description{
Define a range parameter
}
\examples{
range_parameter(
  id = "ks",
  lower_default = 3,
  upper_default = 15,
  lower_distribution = uniform_distribution(1, 5),
  upper_distribution = uniform_distribution(10, 20),
  description = "The numbers of clusters to be evaluated."
)
}
