% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_functions.R
\name{as_paramhelper}
\alias{as_paramhelper}
\alias{as_roxygen}
\alias{as_descriptive_tibble}
\title{Common transformation functions for dynparam objects}
\usage{
as_paramhelper(x)

as_roxygen(x)

as_descriptive_tibble(x)
}
\arguments{
\item{x}{A dynparam object (e.g. distribution, parameter, parameter_set)}
}
\description{
Common transformation functions for dynparam objects
}
