% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_super.R
\name{distribution}
\alias{distribution}
\alias{distribution_function}
\alias{quantile_function}
\alias{print.distribution}
\alias{as.list.distribution}
\alias{as_distribution}
\alias{is_distribution}
\title{Helper functions for converting distributions from and to other formats}
\usage{
distribution(...)

distribution_function(dist)

quantile_function(dist)

\method{print}{distribution}(x, ...)

\method{as.list}{distribution}(x, ...)

as_distribution(li)

is_distribution(x)
}
\arguments{
\item{...}{Fields to be saved in the distribution.}

\item{dist}{A distribution object.}

\item{x}{An object which might be a distribution.}

\item{li}{A list to be converted into a distribution.}
}
\description{
Helper functions for converting distributions from and to other formats
}
\seealso{
\code{\link[=uniform_distribution]{uniform_distribution()}}, \code{\link[=normal_distribution]{normal_distribution()}}, \code{\link[=expuniform_distribution]{expuniform_distribution()}}, \link{dynparam}
}
