% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_expuniform.R
\name{expuniform_distribution}
\alias{expuniform_distribution}
\title{Exponentially scaled uniform distribution.}
\usage{
expuniform_distribution(lower, upper)
}
\arguments{
\item{lower}{Lower limit of the distribution.}

\item{upper}{Upper limit of the distribution.}
}
\description{
Exponentially scaled uniform distribution.
}
\examples{
expuniform_distribution(1, 10000)

expuniform_distribution(1e-5, 1e-2)
}
