% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_logical.R
\name{logical_parameter}
\alias{logical_parameter}
\title{Define a logical parameter}
\usage{
logical_parameter(id, default, description = NULL, length = 1)
}
\arguments{
\item{id}{The name of the parameter.}

\item{default}{The default value of the parameter.}

\item{description}{An optional (but recommended) description of the parameter.}

\item{length}{The length of the vector of this parameter (default 1).}
}
\description{
Define a logical parameter
}
\examples{
logical_parameter(
  id = "scale",
  default = TRUE,
  description = "Whether or not to scale the input variables"
)
}
