% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_super.R, R/set.R
\name{parameter}
\alias{parameter}
\alias{as.character.parameter}
\alias{as.list.parameter}
\alias{as_parameter}
\alias{is_parameter}
\alias{print.parameter}
\alias{as_roxygen.parameter}
\alias{as_parameter_set}
\alias{as.character.parameter_set}
\alias{print.parameter_set}
\title{Helper functions for converting parameters from and to other formats}
\usage{
parameter(id, default, ..., description = NULL, tuneable = TRUE)

\method{as.character}{parameter}(x, ...)

\method{as.list}{parameter}(x, ...)

as_parameter(li)

is_parameter(x)

\method{print}{parameter}(x, ...)

\method{as_roxygen}{parameter}(x)

as_parameter_set(li)

\method{as.character}{parameter_set}(x, ...)

\method{print}{parameter_set}(x, ...)
}
\arguments{
\item{id}{The name of the parameter.}

\item{default}{The default value of the parameter.}

\item{...}{Extra fields to be saved in the parameter.}

\item{description}{An optional (but recommended) description of the parameter.}

\item{tuneable}{Whether or not a parameter is tuneable.}

\item{x}{An object (parameter or distribution) to be converted.}

\item{li}{A list to be converted into a parameter.}
}
\description{
Helper functions for converting parameters from and to other formats
}
\seealso{
\code{\link[=character_parameter]{character_parameter()}}, \code{\link[=integer_parameter]{integer_parameter()}}, \code{\link[=logical_parameter]{logical_parameter()}}, \code{\link[=numeric_parameter]{numeric_parameter()}}, \code{\link[=integer_range_parameter]{integer_range_parameter()}}, \code{\link[=numeric_range_parameter]{numeric_range_parameter()}}, \code{\link[=subset_parameter]{subset_parameter()}}, \link{dynparam}
}
