% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{parameter_set}
\alias{parameter_set}
\alias{is_parameter_set}
\alias{as_paramhelper.parameter_set}
\alias{as.list.parameter_set}
\title{Parameter set helper functions}
\usage{
parameter_set(..., parameters = NULL, forbidden = NULL)

is_parameter_set(x)

\method{as_paramhelper}{parameter_set}(x)

\method{as.list}{parameter_set}(x, ...)
}
\arguments{
\item{...}{Parameters to wrap in a parameter set.}

\item{parameters}{A list of parameters to wrap in a parameter set.}

\item{forbidden}{States forbidden region of parameter via a character vector, which will be turned into an expression.}

\item{x}{An object for which to check whether it is a parameter set.}
}
\description{
Parameter set helper functions
}
\seealso{
\code{\link[=character_parameter]{character_parameter()}}, \code{\link[=integer_parameter]{integer_parameter()}}, \code{\link[=logical_parameter]{logical_parameter()}}, \code{\link[=numeric_parameter]{numeric_parameter()}}, \code{\link[=range_parameter]{range_parameter()}}, \code{\link[=subset_parameter]{subset_parameter()}}, \link{dynparam}
}
