% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_super.R, R/param_super.R
\name{distribution_function}
\alias{distribution_function}
\alias{quantile_function}
\alias{list_as_distribution}
\alias{as_paramhelper}
\alias{as_roxygen}
\alias{as_argparse}
\alias{as_list}
\alias{list_as_parameter}
\title{Helper functions for converting parameters from and to other formats}
\usage{
distribution_function(param)

quantile_function(param)

list_as_distribution(li)

as_paramhelper(param)

as_roxygen(param)

as_argparse(param)

as_list(param)

list_as_parameter(li)
}
\arguments{
\item{param}{The parameter to be converted.}

\item{li}{A list to be converted into a parameter.}
}
\description{
Helper functions for converting parameters from and to other formats
}
