% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_range.R
\name{range_parameter}
\alias{range_parameter}
\title{Define a range parameter}
\usage{
range_parameter(
  id,
  default,
  lower_distribution,
  upper_distribution,
  description = NULL,
  tuneable = TRUE
)
}
\arguments{
\item{id}{The name of the parameter.}

\item{default}{The default value of the parameter.}

\item{lower_distribution}{A distribution from which the lower value of the range can be sampled.}

\item{upper_distribution}{A distribution from which the upper value fo the range can be sampled.}

\item{description}{An optional (but recommended) description of the parameter.}

\item{tuneable}{Whether or not a parameter is tuneable.}
}
\description{
Define a range parameter
}
