% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_range.R
\name{range_parameter}
\alias{range_parameter}
\title{Define a range parameter}
\usage{
range_parameter(id, default, lower_distribution, upper_distribution,
  as_integer = TRUE, description = NULL, tuneable = TRUE)
}
\arguments{
\item{id}{The name of the parameter.}

\item{default}{The default value of the parameter.}

\item{lower_distribution}{A distribution from which the lower value of the range can be sampled.
See \link[dynparam:dynparam]{?dynparam} for a list of possible distributions.}

\item{upper_distribution}{A distribution from which the upper value fo the range can be sampled.
See \link[dynparam:dynparam]{?dynparam} for a list of possible distributions.}

\item{as_integer}{Whether or not this parameter should be interpreted as an integer.}

\item{description}{An optional (but recommended) description of the parameter.}

\item{tuneable}{Whether or not a parameter is tuneable.}
}
\description{
Define a range parameter
}
\examples{
range_parameter(
  id = "ks",
  as_integer = TRUE,
  default = c(3L, 15L),
  lower_distribution = uniform_distribution(1L, 5L),
  upper_distribution = uniform_distribution(10L, 20L),
  description = "The numbers of clusters to be evaluated."
)

range_parameter(
  id = "quantiles",
  as_integer = FALSE,
  default = c(0.1, 0.99),
  lower_distribution = uniform_distribution(0, 0.25),
  upper_distribution = uniform_distribution(0.9, 1),
  description = "The lower and upper quantile thresholds."
)
}
