% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_range.R
\name{range_parameter}
\alias{range_parameter}
\title{Define a range parameter}
\usage{
range_parameter(id, default, lower_distribution, upper_distribution,
  description = NULL)
}
\arguments{
\item{id}{The name of the parameter.}

\item{default}{The default value of the parameter.}

\item{lower_distribution}{A distribution from which the lower value of the range can be sampled.
See \link[dynparam:dynparam]{?dynparam} for a list of possible distributions.}

\item{upper_distribution}{A distribution from which the upper value fo the range can be sampled.
See \link[dynparam:dynparam]{?dynparam} for a list of possible distributions.}

\item{description}{An optional (but recommended) description of the parameter.}
}
\description{
Define a range parameter
}
\examples{
range_parameter(
  id = "ks",
  default = c(3L, 15L),
  lower_distribution = uniform_distribution(1L, 5L),
  upper_distribution = uniform_distribution(10L, 20L),
  description = "The numbers of clusters to be evaluated."
)
}
