% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R, R/super_functions.R
\name{sip.parameter_set}
\alias{sip.parameter_set}
\alias{sip}
\title{It's like \code{sample()}, but for dynparam}
\usage{
\method{sip}{parameter_set}(x, n = 1, as_tibble = TRUE)

sip(x, n = 1, as_tibble = TRUE)
}
\arguments{
\item{x}{A dynparam object (e.g. distribution, parameter, parameter_set)}

\item{n}{Number of objects to return}

\item{as_tibble}{Whether or not to return as a tibble}
}
\description{
It's like \code{sample()}, but for dynparam
}
