% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_character.R, R/param_integer.R,
%   R/param_logical.R, R/param_numeric.R, R/param_range.R, R/param_subset.R,
%   R/super_functions.R
\name{as_descriptive_tibble.character_parameter}
\alias{as_descriptive_tibble.character_parameter}
\alias{as_descriptive_tibble.integer_parameter}
\alias{as_descriptive_tibble.logical_parameter}
\alias{as_descriptive_tibble.numeric_parameter}
\alias{as_descriptive_tibble.range_parameter}
\alias{as_descriptive_tibble.subset_parameter}
\alias{as_paramhelper}
\alias{as_roxygen}
\alias{as_descriptive_tibble}
\alias{sip}
\title{Common transformation functions for dynparam objects}
\usage{
\method{as_descriptive_tibble}{character_parameter}(x)

\method{as_descriptive_tibble}{integer_parameter}(x)

\method{as_descriptive_tibble}{logical_parameter}(x)

\method{as_descriptive_tibble}{numeric_parameter}(x)

\method{as_descriptive_tibble}{range_parameter}(x)

\method{as_descriptive_tibble}{subset_parameter}(x)

as_paramhelper(x)

as_roxygen(x)

as_descriptive_tibble(x)

sip(x)
}
\arguments{
\item{x}{A dynparam object (e.g. distribution, parameter, parameter_set)}
}
\description{
Common transformation functions for dynparam objects
}
