% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_uniform.R
\name{uniform_distribution}
\alias{uniform_distribution}
\title{Uniform distribution}
\usage{
uniform_distribution(lower, upper)
}
\arguments{
\item{lower}{Lower limit of the distribution.}

\item{upper}{Upper limit of the distribution.}
}
\description{
Uniform distribution
}
\examples{
uniform_distribution(1, 10)
}
